UI: Add Dialogue component, testing libquest
The Dialogue component can be added to the sidebar hole of the
QuestFTHView component.

Also, add a UI demo which uses the new component above, and
libquest. This demo plays a small quest, in which the user has to flip
the app, toggle two switches, then flip back to complete the quest.

https://phabricator.endlessm.com/T29384