toolbox: Side by side toolbox
This patch adds a new property for the QuestFTHView to show the toolbox
on the back or side by side.

If sideBySide is set the new component SlideToHack is used instead of
the FlipToHack. This new component just shows the toolbox using the half
of the screen and moves the app to the right.

https://phabricator.endlessm.com/T29689