CI: Use Dockerfile for container build
Rather than running the whole build and root and therefore creating
files in build that can't be deleted by jenkins, provide a build
Dockerfile that installs the packages and use that as the container for
the build. The .dockerignore file prevents an existing build directory
from being sent to the docker daemon only to be ignored.

https://phabricator.endlessm.com/T29701