CI: Put NPM cache in workspace
Inside the container $HOME is /, so the default NPM cache location is
/.npm. That's not writable by the unprivileged jenkins user and it would
be difficult to pre-create it in the container image since we don't know
the UID of the jenkins user beforehand. Instead, set the
NPM_CONFIG_CACHE environment variable to set the location of the cache
inside the workspace. While here, disable NPM update check notifications
since there's no way to do it as the unprivileged user, anyways.

https://phabricator.endlessm.com/T29701