CI: Allow flatpak to run from the docker container
Both docker and flatpak (via bubblewrap) implement sandboxes. Bubblewrap
will fail if it doesn't have the privileges to do that, so various holes
need to be poked in docker's sandbox. Furthermore, flatpak will run
tools such as ldconfig that lookup the running user in the databases.
Since we want to run as the host jenkins user but that's obviously not
in the container, bind mount the host's passwd and group files read
only.

https://phabricator.endlessm.com/T29701