Sidetrack: Expose game variables to quest
This patch expose some variables to the sidetrack quest and links the
changes between the quest variables and the app. If a variable is
changed in the quest, the app will be updated and the same thing will
happen in the other direction.

This will allow us to create different quests with the same app, just
updating that variables and waiting for them to change.

Note:
This change depends on a change in the hack-toy-apps so it's required to
run: git submodule update --remote

https://phabricator.endlessm.com/T29826