Sidetrack: Get quest params only when quest changes
The useEffect with the quest creates a race condition when the level
ends, the app updates the quest and the quest updates the app and that
causes some problems if a variable is updated in the app.

To solve this problem this patch checks if the quest has really changed,
so this code will only be called when the quest state changes.

https://phabricator.endlessm.com/T29906