FTH: Add attract state for flip to hack button
This patch adds the attract state to the FTHButton component and also
adds a new variable to the sidetrack quest to enable/disable this new
attract state just modifying the attractFTH variable. The flipped
variable is also exposed to sidetrack to be able to detect when the app
was flipped.

https://phabricator.endlessm.com/T29916