toolbox: Fix code editor delay
The code editor in the toolbox modifies the app and when the app is
modified the code is updated causing some issues when the code changes
during this process.

This patch fixes this issue just removing the update of the code in the
p5 and html quests, we only need the update of the app using the code
and we don't need the other direction update.

https://phabricator.endlessm.com/T29927