sidetrack: remove timeout initialization
This patch changes how the sidetrack app is initialized improving how it
loads the iframe in the useEffect by avoiding non needed calls.

The useEffect to initialize the sidetrack app is called only once, using
`useRef(quest)` instead of the quest variable directly inside the
effect.

The game initialization is done now in the loadState, instead of do the
loadState and later update the startLevel. This changes a bit the
behavior of variables in the ink quest because the initial level will be
the highestAchievedLevel, and to change the level during the quest the
startLevel variable should be updated. The currentLevel is readOnly, so
modifying that in the quest won't change the level.

https://phabricator.endlessm.com/T29944