sidetrack: Fix level reloading
The app restart cannot be done just modifying the startLevel variable
because the sidetrack app just check for changes and the startLevel
doesn't change if we don't change by hand, so we need to enforce the
scene reload.

This will fix the problem of the first level twice because it was setted
by the quest variable modification callback that is called when the
dialog changes and for the first time it was getting the first
startLevel assignement.

https://phabricator.endlessm.com/T29987