UI: Change drawer width according to breakpoints
The drawer width is hardcoded to 3 of 12 columns of the corresponding
breakpoints: md, lg and xl. Note: The value is hardcoded instead of
fluid, because otherwise it would reflow the text inside the scroller
view. And also because this is component is not actually part of the
main grid: is displaced in and out of the screen.

https://phabricator.endlessm.com/T30001