deploy: Different bucket depending on HOME
The deployment using the jenkins job is unable to get the git branch
name because how the checkout is done so we shouldn't use that to get
the branch name.

This patch uses the HOME variable instead that is hack-web-master or
hack-web-stable during the CI process. It will deploy to
dev.hack-computer.com or try.hack-computer.com AWS S3 bucket depending
on that so we don't need commits in stable that are not in master.

https://phabricator.endlessm.com/T30017