toolbox: Fix sidetrack code error in chrome
The way to detect the line and column of an error in chrome is different
so we need to update the regular expression.

This patch also makes this code more robust because in case that a line
is not found the error will be shown at the top, but won't crash.

https://phabricator.endlessm.com/T30031