Copy ace-editor workers
The ace-editor has some workers that provides some code validation
depending on the mode, this patch adds a needed import to fix that.

https://phabricator.endlessm.com/T29998

This patch also provides a better editor code validation done by default
by the editor itself, so we've a basic javascript and basic html
validation that we were not using before.

https://phabricator.endlessm.com/T29989