toolbox: Fix code editor build annotations
The custom annotations in the code editor was not working correctly
because it's a state property and if the annotation is setted with the
same object, the component doesn't trigger the render and the
annotations dissapears.

This patch fix the problem with custom annotations for sidetrack just
adding a unique id to each annotation created so we can ensure that is
different from the previous one.

This patch also improves the build method using the debounceChangePeriod
property from AceEditor instead of a custom timeout.

https://phabricator.endlessm.com/T30044