sidetrack: Only update toolbox state on level change
Updating the toolbox state with every app change introduces a problem in
the code editor that shows the previous state if the user types just
between an update process.

We don't need to refresh the toolbox state with every change, we only
need to update the state when the level changes.

https://phabricator.endlessm.com/T30064