toolbox: Fix code editor update while editing
The build function can take some time, if the text is edited during the
build we can have inconsistences because the text in the editor changes
but after the build the text changes back again.

This patch convert the text to a component state variable and updates
that on every change. The build is launched with a timeout so the text
is updated for the user instantly removing the effect of code appearing
or dissapearing when the build finish.

https://phabricator.endlessm.com/T30064