Update the document.title to match the topbar title
The title give some information to the user and also to the google
analytics. This patch modifies the HackTopBar component to update the
document.title with the topbar title so we'll have the complete title in
the browser.

As a side effect we'll receive more information in the analytics,
because this title will be attached with the path and it's more
readable.

https://phabricator.endlessm.com/T30149