Add typescript-eslint/parser dev dependency
There's a warning in the build telling us that the peer dependency is
not installed.

npm WARN @typescript-eslint/eslint-plugin@2.31.0 requires a peer of @typescript-eslint/parser@^2.0.0 but none is installed. You must install peer dependencies yourself.

This patch adds the typescript-eslint/parser package to the dev
dependencies to solve this problem.

https://phabricator.endlessm.com/T29832