MainButton: Wrap buttons with AttractWrapper
This patch adds a new wrap class around all the buttons defined in
main-button. This wrapper just adds a new property `attracting` and sets
the `glow` css class if the `attracting` property is true.

This will help us to attract the user attention to our buttons just by
changing this new property.

https://phabricator.endlessm.com/T30196