libquest: Add choice modifiers
This patch adds some choice modifiers to libquest to be able to
customize some options in the interface.

Right now there's only the attracting modifiers, but it's implemented
in a general way so we can add more modifiers in the future, for
example to change the color, add other animations, etc.

https://phabricator.endlessm.com/T30196