ui: Fix useCard store selector with empty state
With the translations change we made the cardset state empty by default
and it's filled just after the i18n is initialized. This introduces a
problem in direct routing to quests because the useCard fails, the
cardset is undefined and we can't access to the cards attribute.

This patch just adds a check for that case and if the cardset is
undefined it returns a default empty card state. This will fix the
direct quest loading for example:

http://localhost:5000/art

https://phabricator.endlessm.com/T30221