libquest: game state external functions
This patch adds two external functions for quests so every ink quest can
use:

get_game_state(key)
set_game_state(key, value)

This will work like a key-value storage, currently stored in a store
reducer, but in a near future we can also persist this data using
localStorage or even a web service.

It's also possible to declare nested keys using a '/', for example the
key 'quest.Web/complete' will be stored like:

{
  quest.Web: {
    complete: value
  }
}

This patch also defines a new function in the store,
trackGameStateChanges, that's called everytime the game state is
changed. This function is used to send metrics to google analytics.

Right now, only a few keys are tracked:

 * 'quest.Web/complete'
 * 'quest.P5/complete'
 * 'quest.Sidetrack1/complete'
 * 'quest.Sidetrack2/complete'

Any other key that it's wanted to be tracked should be added to that
function.

https://phabricator.endlessm.com/T30256