sidetrack: Fix restartApp
Restart should update the startLevel, in other case, it's not possible
to go back to the previous startLevel in the future.

This commit also resets the instructionCode and levelCode for each level
on restartApp so if we reach the same level we'll get the initial state
and not the solved state.

https://phabricator.endlessm.com/T30364