dialog: Add custom class for achievements messages
Achievement messages have an image in the html code. This image takes
some time to render and sometimes the scroll is done before the image is
completely loaded, causing the scroll problem.

To fix this we can set a fixed width/height, so the `img` element always
has the same size and the browser will scroll correctly even if the img
loading takes some time.

This patch adds a new class for achievements messages and sets a fixed
width and height.

It's done this way, instead of applying the same style for all messages
to avoid forcing this style for possible quest messages with images.

With this change it's also possible to add move custom styles for
different message types, like messages with snippets.

https://phabricator.endlessm.com/T30335