Display a disclaimer for using google analytics
This commit adds a new component, CookieDialog. This new component is
added to the main app so it's there on every "page".

To show only once, this component adds a cookie when it's closed, so it
should appear the first time. To be able to see it again, cookies should
be cleaned.

The Google Analytics is initialized when the user closes the dialog, so
the app doesn't track any information until the user confirms.

https://phabricator.endlessm.com/T30222