Convert cookie disclaimer to a top banner
This patch convert the cookies disclaimer to a top cookies banner.

There are two buttons, one to allow cookies and another one to block.

This could be completed with a more detailed description or maybe a
custom privacy policy page with a detailed description of each cookie
stored in the browser.

https://phabricator.endlessm.com/T30222