sidetrack: Move mute state to the store
There are more components that play sounds, like the flip to hack
button. Right now the mute button in sidetrack just mutes the sidetrack
app, but not the other components.

This commit moves the mute state to the store so other components can
check that and silence sounds accordingly.

As a side effect, the mute state will be kept between quest changes.

https://phabricator.endlessm.com/T30418