html: Add translations to the app
The html code in the app has a lot of english text so if we want to
translate the quest we need a way to translate the html itself.

This patch adds an initialization function to the app that looks for the
language sets so we can load different html or css depending on the
language selected.

This patch also adds an initial spanish translation for this html code.

https://phabricator.endlessm.com/T30413