inventory: Remove minHeight from badges
The min height in the badge component is not responsive. The 100% width
will make this reponsive so it'll scale with the grid size.

The paddingTop css is needed because the badge doesn't have any visible
element inside so it has 0 height, setting the paddingTop to 100% makes
it a rectangle.

padding-top and padding-bottom is based on the parent element's width.
So if we had an element that is 500px wide, and padding-top of 100%,
the padding-top would be 500px.

https://developer.mozilla.org/en-US/docs/Web/CSS/padding-top#Values

https://phabricator.endlessm.com/T30459