Sidetrack quest: Fix saying level complete
In the end_level_check tunnel, it shouldn't say level complete when
the player is already in the desired level. The tunnel should
fast-forward directly to the next step.

In the mid_level_check tunnel, the "level x complete!" was
missing. Instead, an empty tag was used to avoid having the next text
presented as a user answer.

https://phabricator.endlessm.com/T30359